refactor: move product.ts to types
