+++ server/app/types/product.ts
